@@ -1,19 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:revolutionary_solitaire/ui/mainScreen.dart';
-import 'ui/Deck.dart';
-import 'ui/winningScreen.dart';
-import 'ui/mainScreen.dart';
+import 'package:revolutionary_solitaire/ui/main_screen.dart';
+
+import 'ui/deck.dart';
+import 'ui/main_screen.dart';
+import 'ui/winning_screen.dart';
 
 void main() => runApp(SolitareApp());
 
 class SolitareApp extends StatelessWidget {
-
-
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-
     SystemChrome.setPreferredOrientations([
       DeviceOrientation.portraitUp,
       DeviceOrientation.portraitDown,
@@ -26,12 +24,10 @@ class SolitareApp extends StatelessWidget {
       ),
       initialRoute: '/',
       routes: {
-        '/'  : (context) => MainScreen(),
-        '/deck' : (context) => GameBoard(title: '27 Red And Black'),
-        '/win'  : (context) => WinningScreen(),
+        '/': (context) => MainScreen(),
+        '/deck': (context) => const GameBoard(title: '27 Red And Black'),
+        '/win': (context) => WinningScreen(),
       },
-
     );
   }
 }
-
