@@ -1,38 +1,25 @@
 import 'package:flutter/material.dart';
 import 'package:revolutionary_solitaire/model/playing_card.dart';
-import 'cardBuilder.dart';
+import 'card_builder.dart';
 
-
-class CardCell extends StatefulWidget{
-
+class CardCell extends StatefulWidget {
   final PlayingCard card;
- final void Function(List) onValueChanged;
+  final void Function(List<PlayingCard>) onValueChanged;
 
-
-  CardCell({
+  const CardCell({
     @required this.card,
     @required this.onValueChanged,
-
-
-  }
-      );
-   @override
-  _cardCell createState()=> _cardCell();
+  });
+  @override
+  CardCellState createState() => CardCellState();
 }
 
-class _cardCell extends State<CardCell>{
+class CardCellState extends State<CardCell> {
   @override
   Widget build(BuildContext context) {
-    return CardBuilder(playingCard: widget.card,onValueChanged: widget.onValueChanged,);
+    return CardBuilder(
+      playingCard: widget.card,
+      onValueChanged: widget.onValueChanged,
+    );
   }
-
 }
-
-
-
-
-
-
-
-
-
